refactor: Edit recommend place button
